requirements.txt
asgiref==3.5.0
Django==3.2.12
pip==22.0.4
pytz==2022.1
setuptools==62.0.0
sqlparse==0.4.2
typing-extensions==4.1.1
wheel==0.37.1
djangorestframework==3.13.1



requests~=2.27.1